Fix monthly income calculation

+++ src-core/src/models.rs
@@ -533,10 +533,10 @@ impl IncomeSummary {
         self.total_income += converted_amount;
     }
 
-    pub fn calculate_monthly_average(&mut self, num_months: Option<f64>) {
-        let months = num_months.unwrap_or_else(|| self.by_month.len() as f64);
-        if months > 0.0 {
-            self.monthly_average = self.total_income / months;
+    pub fn calculate_monthly_average(&mut self, num_months: Option<u32>) {
+        let months = num_months.unwrap_or_else(|| self.by_month.len() as u32);
+        if months > 0 {
+            self.monthly_average = self.total_income / (months as f64);
         }
     }
 }
